- Changed the data type of const fields (ELEMENT_HEIGHT_UNIT,ELEMENT_WIDTH_UNIT) from float to int. - In the MainMenuScreen class changed the old variables of units to new constants.

+++ core/src/ua/carcassone/game/Utils.java
@@ -4,13 +4,13 @@ import com.badlogic.gdx.Gdx;
 
 public class Utils {
 
-    private static final float SCALING_COEFFICIENT = 12;
+    private static final int SCALING_COEFFICIENT = 12;
 
     /**
      * standard unit for interactive elements on the screen (buttons, labels, etc.).
      * */
-    public static final float ELEMENT_HEIGHT_UNIT = Gdx.graphics.getHeight() / SCALING_COEFFICIENT;
-    public static final float ELEMENT_WIDTH_UNIT = Gdx.graphics.getWidth() / SCALING_COEFFICIENT;
+    public static final int ELEMENT_HEIGHT_UNIT = Gdx.graphics.getHeight() / SCALING_COEFFICIENT;
+    public static final int ELEMENT_WIDTH_UNIT = Gdx.graphics.getWidth() / SCALING_COEFFICIENT;
 
     /**
      * Returns an amount of pixels from the bottom of the screen to match some amount of pixels from the top.
